with_: implement context manager function
This function requires the user to extract the bound value from a
`Maybe` container object, if a value is bound by the context manager.
This is explained in the reference.

Implementing this necessitated using a `Maybe` container for
possibly-missing values because of the way the lambdak running mechanism
works internally.