Represent lambdak stored function args as tuple
If the function doesn't take any arguments, represent that as the empty
tuple. The convenience of this is we can call the lambdak's `k` function
with `*x` and Python will automatically expand that into the correct
number of arguments, or no arguments at all if `x` is an empty tuple.

This lets us simplify the `with_` function's `act_k` handler function
because now we can call it will zero or one argument depending on what's
bound by the context manager.